bug fix in fill up functions

+++ ufo.py
@@ -213,7 +213,7 @@ def rotationMatrix(axis, theta):
                      [2*(bd+ac), 2*(cd-ab), aa+dd-bb-cc]])
                      
 def fillUp(strip, color, ordering, wait_ms = 10):
-    for h in range(min(ordering),max(ordering),(max(ordering)-min(ordering)/100.0)):
+    for h in range(int(min(ordering)),int(max(ordering)),int((max(ordering)-min(ordering)/100.0))):
         for p in range(924):
             if ordering[p] < h:
                 strip.setPixelColor(p, color)
@@ -223,7 +223,7 @@ def fillUp(strip, color, ordering, wait_ms = 10):
         time.sleep(wait_ms/1000.0)
         
 def fillUpFire(strip, ordering, wait_ms=10):
-    for h in range(min(ordering),max(ordering),(max(ordering)-min(ordering)/100.0)):
+    for h in range(int(min(ordering)),int(max(ordering)),int((max(ordering)-min(ordering)/100.0))):
         for p in range(924):
             strip.setPixelColor(p, fireColor((h-ordering[p])*1500/(max(ordering)-min(ordering))))
         strip.show()
